# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface User {
  id: string;
  deviceId: string;
  createdAt: string;
}

export interface DeviceAuthRequest {
  /** Stable device identifier generated by the client. */
  deviceId: string;
}

export interface DeviceAuthResponse {
  /** Bearer access token (JWT). */
  token: string;
  user: User;
}

export interface Photo {
  id: string;
  fortId: string;
  landmarkId?: string | null;
  caption?: string | null;
  tags: string[];
  /** Public URL to the image file. */
  imageUrl: string;
  userId: string;
  createdAt: string;
}

export interface CreatePhotoRequest {
  image: Blob;
  landmarkId?: string;
  caption?: string;
  /** Comma-separated tags (e.g. \"gate, sunrise, trek\") */
  tags?: string;
}

export interface PhotoListResponse {
  items: Photo[];
  nextCursor: string | null;
}

export interface DeletePhotoResponse {
  ok: boolean;
}

export type ListFortPhotosParams = {
  tag?: string;
  landmarkId?: string;
  /**
   * @minimum 1
   * @maximum 100
   */
  limit?: number;
  /**
   * Opaque pagination cursor
   */
  cursor?: string;
};