# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: auth
    description: Authentication operations
  - name: photos
    description: Community photo operations
paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      description: Returns server health status
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"
  /auth/device:
    post:
      operationId: authDevice
      tags: [auth]
      summary: Device authentication
      description: Creates (or finds) a device-bound user and returns an access token.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DeviceAuthRequest"
      responses:
        "200":
          description: Authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DeviceAuthResponse"
  /forts/{fortId}/photos:
    get:
      operationId: listFortPhotos
      tags: [photos]
      summary: List community photos for a fort
      parameters:
        - name: fortId
          in: path
          required: true
          schema: { type: string }
        - name: tag
          in: query
          required: false
          schema: { type: string }
        - name: landmarkId
          in: query
          required: false
          schema: { type: string }
        - name: limit
          in: query
          required: false
          schema: { type: integer, minimum: 1, maximum: 100, default: 50 }
        - name: cursor
          in: query
          required: false
          schema: { type: string, description: Opaque pagination cursor }
      responses:
        "200":
          description: Photo list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PhotoListResponse"
    post:
      operationId: createFortPhoto
      tags: [photos]
      summary: Upload a community photo for a fort
      security:
        - bearerAuth: []
      parameters:
        - name: fortId
          in: path
          required: true
          schema: { type: string }
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              $ref: "#/components/schemas/CreatePhotoRequest"
      responses:
        "200":
          description: Created photo
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Photo"
  /photos/{photoId}:
    get:
      operationId: getPhoto
      tags: [photos]
      summary: Get photo metadata
      parameters:
        - name: photoId
          in: path
          required: true
          schema: { type: string, format: uuid }
      responses:
        "200":
          description: Photo
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Photo"
    delete:
      operationId: deletePhoto
      tags: [photos]
      summary: Delete a photo (owner only)
      security:
        - bearerAuth: []
      parameters:
        - name: photoId
          in: path
          required: true
          schema: { type: string, format: uuid }
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DeletePhotoResponse"
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required:
        - status
    User:
      type: object
      properties:
        id:
          type: string
          format: uuid
        deviceId:
          type: string
        createdAt:
          type: string
          format: date-time
      required: [id, deviceId, createdAt]
    DeviceAuthRequest:
      type: object
      properties:
        deviceId:
          type: string
          description: Stable device identifier generated by the client.
      required: [deviceId]
    DeviceAuthResponse:
      type: object
      properties:
        token:
          type: string
          description: Bearer access token (JWT).
        user:
          $ref: "#/components/schemas/User"
      required: [token, user]
    Photo:
      type: object
      properties:
        id:
          type: string
          format: uuid
        fortId:
          type: string
        landmarkId:
          type: string
          nullable: true
        caption:
          type: string
          nullable: true
        tags:
          type: array
          items: { type: string }
        imageUrl:
          type: string
          description: Public URL to the image file.
        userId:
          type: string
          format: uuid
        createdAt:
          type: string
          format: date-time
      required: [id, fortId, tags, imageUrl, userId, createdAt]
    CreatePhotoRequest:
      type: object
      properties:
        image:
          type: string
          format: binary
        landmarkId:
          type: string
        caption:
          type: string
        tags:
          type: string
          description: Comma-separated tags (e.g. \"gate, sunrise, trek\")
      required: [image]
    PhotoListResponse:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/Photo"
        nextCursor:
          type: string
          nullable: true
      required: [items, nextCursor]
    DeletePhotoResponse:
      type: object
      properties:
        ok:
          type: boolean
      required: [ok]