# lib/api-zod/src/generated/api.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import * as zod from "zod";

/**
 * Returns server health status
 * @summary Health check
 */
export const HealthCheckResponse = zod.object({
  status: zod.string(),
});

/**
 * Creates (or finds) a device-bound user and returns an access token.
 * @summary Device authentication
 */
export const AuthDeviceBody = zod.object({
  deviceId: zod
    .string()
    .describe("Stable device identifier generated by the client."),
});

export const AuthDeviceResponse = zod.object({
  token: zod.string().describe("Bearer access token (JWT)."),
  user: zod.object({
    id: zod.string().uuid(),
    deviceId: zod.string(),
    createdAt: zod.coerce.date(),
  }),
});

/**
 * @summary List community photos for a fort
 */
export const ListFortPhotosParams = zod.object({
  fortId: zod.coerce.string(),
});

export const listFortPhotosQueryLimitDefault = 50;
export const listFortPhotosQueryLimitMax = 100;

export const ListFortPhotosQueryParams = zod.object({
  tag: zod.coerce.string().optional(),
  landmarkId: zod.coerce.string().optional(),
  limit: zod.coerce
    .number()
    .min(1)
    .max(listFortPhotosQueryLimitMax)
    .default(listFortPhotosQueryLimitDefault),
  cursor: zod.coerce.string().optional(),
});

export const ListFortPhotosResponse = zod.object({
  items: zod.array(
    zod.object({
      id: zod.string().uuid(),
      fortId: zod.string(),
      landmarkId: zod.string().nullish(),
      caption: zod.string().nullish(),
      tags: zod.array(zod.string()),
      imageUrl: zod.string().describe("Public URL to the image file."),
      userId: zod.string().uuid(),
      createdAt: zod.coerce.date(),
    }),
  ),
  nextCursor: zod.string().nullable(),
});

/**
 * @summary Upload a community photo for a fort
 */
export const CreateFortPhotoParams = zod.object({
  fortId: zod.coerce.string(),
});

export const CreateFortPhotoBody = zod.object({
  image: zod.instanceof(File),
  landmarkId: zod.string().optional(),
  caption: zod.string().optional(),
  tags: zod
    .string()
    .optional()
    .describe('Comma-separated tags (e.g. \\\"gate, sunrise, trek\\\")'),
});

export const CreateFortPhotoResponse = zod.object({
  id: zod.string().uuid(),
  fortId: zod.string(),
  landmarkId: zod.string().nullish(),
  caption: zod.string().nullish(),
  tags: zod.array(zod.string()),
  imageUrl: zod.string().describe("Public URL to the image file."),
  userId: zod.string().uuid(),
  createdAt: zod.coerce.date(),
});

/**
 * @summary Get photo metadata
 */
export const GetPhotoParams = zod.object({
  photoId: zod.coerce.string().uuid(),
});

export const GetPhotoResponse = zod.object({
  id: zod.string().uuid(),
  fortId: zod.string(),
  landmarkId: zod.string().nullish(),
  caption: zod.string().nullish(),
  tags: zod.array(zod.string()),
  imageUrl: zod.string().describe("Public URL to the image file."),
  userId: zod.string().uuid(),
  createdAt: zod.coerce.date(),
});

/**
 * @summary Delete a photo (owner only)
 */
export const DeletePhotoParams = zod.object({
  photoId: zod.coerce.string().uuid(),
});

export const DeletePhotoResponse = zod.object({
  ok: zod.boolean(),
});