# lib/api-zod/src/generated/types/createPhotoRequest.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface CreatePhotoRequest {
  image: Blob;
  landmarkId?: string;
  caption?: string;
  /** Comma-separated tags (e.g. \"gate, sunrise, trek\") */
  tags?: string;
}