# lib/api-zod/src/generated/types/deviceAuthRequest.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface DeviceAuthRequest {
  /** Stable device identifier generated by the client. */
  deviceId: string;
}