# lib/api-zod/src/generated/types/photo.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface Photo {
  id: string;
  fortId: string;
  landmarkId?: string | null;
  caption?: string | null;
  tags: string[];
  /** Public URL to the image file. */
  imageUrl: string;
  userId: string;
  createdAt: Date;
}