# lib/api-zod/src/generated/types/photoListResponse.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { Photo } from "./photo";

export interface PhotoListResponse {
  items: Photo[];
  nextCursor: string | null;
}